Implemented Collisions and Triggers
Added a Circle Collider 2D component to the three spike ball game objects.

Added a Box Collider 2D component to the Player game object. Added a RigidBody 2D component and set it to Dynamic (so that it can detect collisions with all other game objects).

Created a Game/Level finish game object. Added a Polygon Collider 2D component to it and set it as a Trigger.

Made changes to the Player C# script. Defined functions for Player Death (which resets player position to start point) and Game Finish/Player Escaped ( which for now just prints a statement). Called the Player Death funtion when player collides with Spike Balls and called the Player Escaped function when player overlaps with Game Finish object.