Created a new Level and Player Death
Created new Scene and named it Level 2. Changed the design of the new level by adding obstacle wall which has same effect on the player as the spike balls and by changing the positioning and number of spike balls. Added the scene in the build settings.

Created game object for LevelManager. Created LevelManager C# script. Created function to load the next level when player reaches the Level Finish game object. Created a function to Restart/Reload the current level when Player comes in contact with a spike ball or obstacle wall (Player Death).