Fixed the sound effects error
The sound effects were not playing after scene reload. This is because of using the PlayOneShot(clip) function in the Sound Manager script for playing sound effects. To solve this sound bug, Play() function was used instead and looping property was disabled in unity inspector for sound effect game object.